Update README.md
Object of the research: information storage, transmission and processing systems. Subject of the research: approaches to implementing the distributed ledger technology, in particular, the blockchain technology, as well as implementation issues and feasibility. Purpose of the work: enhancing the safety of data storage, transmission and processing by embedding distributed ledgers in the existing information system. Research methods: the theoretical and analytical research is supposed to underlie the study of the experience building upon the current application and possible options for boosting the potential of the distributed ledger technology, in particular, the blockchain technology. Scientific novelty. The results of the research are supposed to be used to issue proposals and recommendations on how to improve the security of data storage and processing through the use of distributed ledger technology. Practical importance. Practical perspectives consist in the fact that the discussed process of developing and implementing a distributed system simplifies further integration of the technology under consideration with the resulting blockchain service enabling data storage inside a non-modifiable registry.